Enable testing.
Some initial tests for important code used by the proxy. These tests don't really test the proxy itself, per se, though. The sitecues URL module would ideally be a standalone project. However, its heavy knowledge of the dev environment makes me wary of open sourcing it. And Ai Squared / sitecues does not currently subscribe to private npm or maintain a private registry, making it hard to list as an abstracted dependency. Therefor, it sems like a good place to start for tests, so it remains an obvious and isolated sub-unit of functionality.