Add TODO about loader settings.
Loader creation just got a big re-design to be independent and generic. Now we need to figure out how to take loader settings, since it is no longer a flat object and most proxy usage is at the command line. Would it ever be acceptable for a user to explicitly provide the entire string? The server might need to know the Site ID for other purposes, which would then demand robust parsing, adding more processing complexity.