Remove configuration responsibility from Server.js
The "front door" FriendlyServer is now responsible for all of this, so that the main Server code can be ignorant about pesky envirionment details. In addition, in cases where hundreds or thousands of proxies need to be launched in quick succession and we are confident about our settings, such as automated testing or on-demand Try-It-Now servers, a large chunk of logic can be safely bypassed. See last few commits for the refactoring and more details.