Begin implementing state overrides in .listen()
Ideally, .listen() would extract, sanitize, permute, and then save all data from its input to the server's state object in exactly the same manner as happened during server construction, but only with the properties relevant to listening. This is a step towards that, by having the listen options delegate to the server's state.