Fix .listen() configuration.
Node's underlying API has bad semantics when using an object to configure .listen(), it calls the hostname part "host'" even though it is incapable of parsing the port from that property. Therefor, we have to either play some musical chairs with the state object (which implements the proper semantics and includes the port in "host", which causes an exception here) or we use this pattern, which gaurantees seperation of hostname and port.