Improve cloud proxy reliability.
Ensure that proxies are only reloaded if `npm install` succeeds, because if it does not, the dependency tree is probably in bad shape and what is in memory is more likely to be reliable. Also workaround the lack of env specific app targets in pm2 deployment and ensure forward proxies are not started in the cloud due to security reasons.