Fixed loaderOption's object and added seth's fix to reorder how we iterate over config object.
SiteID sometimes would come after we used it as iterating over an object like config is not a guaranteed order, so we made a sorting solution to guarantee siteID's order.

Fixed the creation of the config object, we were initially trying to add siteID without defining the config object's structure.