Remove lots of unnecessary computation.
Buried in the page rendering algorithms was a hack that was eating up quite a bit of CPU time, left over from an era when our URL resolution was not as robust. In addition to performance, it was actively sabotaging how we intended to resolve anchor hrefs in some cases. Complex pages now render more reliably and noticeably faster.