Canonicalize relative URLs via redirects.
This fixes a weird footgun case. We would successfully retrieve a stylesheet that came in via a relative URL, because we intelligently resolved it based on its referrer header. But then that content contained a relative URL, causing the client to make another request, only this time the referrer pointed to the relative URL we resolved earlier, meaning we no longer had any absolute target to infer. And thus the second request was unresolvable. Oh my.