Ship the proxy CA with its private key.
Warning: This CA is for development purposes only, do not use it for anything public-facing. The private key should be considered insecure due to being online in version control and unecrypted. It is here to make development easier. To provide some level of protection, the proxy CA is an intermediate that can be revoked or recreated by the root CA when necessary. There be dragons.