Simplify architecture and remove Try It. (#2)
* Hapi experiment.

No sitecues injection yet. But most of the infrastructure is there.

* Use zlib streams to handle encoded responses.

This sacrifices CPU for improved network performance, as with most encoding processes. Would be nice to benchmark this at some point.

* Resolve domain root URLs, passthrough non-HTML.

* Deal with edge cases for relative targets.

* Defend against invalid targets.

* Remove obsolete imports.

* Simplify attribute processing and logging.

* Simplify logging.

* Use ES2016 Array#includes to improve readability.

* Sync up target and stream-target routes.

* Breaking up request and response handlers.

* Spacing tweaks.

* Core modules come first.

* Put DOM modification in its own module.

* Remove localStatePassthrough, will never use it.

This config option allows cookies intended for the proxy to be passed upstream. For our use case, this would be a security problem. Don't use it.

* Support XHTML.

* SImplify route import and response header mapping.

* Standardize private proxy routes.

Change status route to reflect its private use. Going forward, all such routes will be prefixed by /__sitecues-proxy/ to isolate them from the rest of the internet.

* Implement intelligent loader detection.

Some customers use website technologies that prevent them from adding attributes to elements or even using inline scripts. We now provide a variety of hints to inform the proxy that a loader is present.

* Handle SIGINT

* Document the process to create SSL infrastructure.

* Cleanup project metadata.

* Tweak legacy code.

* Create proper API and redesign CLI to use it.

* Singular directory names.

* Standardize dir name for private keys.

* Add SSL certificates.

* Ship the proxy CA with its private key.

Warning: This CA is for development purposes only, do not use it for anything public-facing. The private key should be considered insecure due to being online in version control and unecrypted. It is here to make development easier. To provide some level of protection, the proxy CA is an intermediate that can be revoked or recreated by the root CA when necessary. There be dragons.

* Enable HTTPS proxying.

* Use filter() for improved readability.

* Wire up the Try It for local HTTPS

Also fix missing dependency.

* Fix lint errors.

* Add Tidy lint automation and fix lint errors.

* Add Tidy lint automation and fix lint errors.

* Remove obsolete cruft: Grunt, JSCS, JSHint.

* Remove obsolete cruft: sanitizePort and test lists

* Fix lint errors.

* Bump deps and fix lint errors.

* Fix lint errors.

* Add Ci configuration.

* Unpin and bump deps. Solve some lint errors.

* Fix and simplify response handler.

* Fix intercept regressions.

* Unpin and bump dependencies.

* Extract response unzipping into new module.

* Extract shutdown code into new handle-quit module.

This is an extremely common pattern for servers. It deserves a life of its own.

* Move the new Try It to its own project.

* Bump dependencies and solve some lint errors.

* Remove obsolete code.

* Tweaks.

* Make the binary executable for ease of use.

* Simplify and remove cruft.

* Allow the user to set the log level.

Also solve some lint errors.

* Ensure the ca is included in package tarballs.

* Solve all lint errors.

* Hack around Hoxy’s bad exit behavior.

Hoxy has a server buried in its internals that it uses to accomplish some neat tricks for serving static files. Unfortunately, there’s no public way to shutdown this server. That conflicts with what we want to do to shutdown gracefully upon receiving SIGINT.

For now, we’re just going to force quit after a reasonable period.

Details: https://github.com/greim/hoxy/issues/88

* General cleanup.

Improve comments, documentation, and error messages. Be smarter about “access denied” errors and don’t suggest to use sudo on Windows.

* Fix error handling for privileged ports.

* Revamp documentation.

* Fix CLI flag conflict.